backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and vector store
supabase==2.3.0
pinecone-client==2.2.4

# LangChain and LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-pinecone==0.0.1
langchain-supabase==0.0.1

# Llama 3.2 integration
llama-cpp-python==0.2.11
transformers==4.36.0
torch==2.1.0

# Embeddings and vector operations
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# HTTP client and async support
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Environment and configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Additional utilities
tiktoken==0.5.1
tenacity==8.2.3
